+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [0.0.3]
 
-1. added CHANGELOG
-2. added 'validate' tags in request/response structures
+1. add CHANGELOG
+2. add 'validate' tags in request/response structures
+3. add validator package & tests